requirements.txt
# Dashboard Financeiro Pessoal - Dependências
# Versões otimizadas e testadas

# Framework principal
streamlit>=1.46.0,<2.0.0

# Manipulação e análise de dados
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Visualização de dados
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Utilitários
python-dateutil>=2.8.0,<3.0.0
pytz>=2023.3

# Desenvolvimento (opcional)
# pytest>=7.0.0,<8.0.0
# black>=23.0.0,<24.0.0
# flake8>=6.0.0,<7.0.0

gspread
google-auth